# Conflicts:
#	Motor2D/FirstScene.cpp
#	Motor2D/FirstScene.h

+++ Motor2D/Grass.h
@@ -14,7 +14,7 @@ public:
 	bool Draw();
 
 private:
-	int CalculateAnimSpeed(float wind)const;
+	float CalculateAnimSpeed(float wind)const;
 
 private:
 	float speed = 0;
